% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{as.tmCorpus}
\alias{as.tmCorpus}
\title{Create textmining Corpus}
\usage{
as.tmCorpus(x, ...)
}
\arguments{
\item{x}{source object, can be tmTaggedText VCorpus or stylo.corpus.}

\item{...}{metadata to be added}
}
\description{
Create textmining Corpus
}
\examples{
tmCorpus(c("This is new corpus", "And I like It"))
as.tmCorpus(c("This is new corpus", "And I like It"))

}

