% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{filter_documents}
\alias{filter_documents}
\title{Function to filter tagged text}
\usage{
filter_documents(x, column, value)
}
\arguments{
\item{x}{tmTaggedCorpus}

\item{column}{column name}

\item{value}{filtered value}
}
\description{
Function to filter tagged text
}
\examples{
\dontrun{
library(dplyr)
library(textmining)
corp <- tmCorpus(c("This is corp corp", "Document 2 corp corp"))
rd <- tmTaggedCorpus(corp)
filtered_tmTaggedCorpus <- filter_documents(rd, "tag", "NP")
corpus <- tmCorpus(filtered_tmTaggedCorpus)
}

}

