% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_functions.R
\name{getDoc}
\alias{getDoc}
\title{Function to access documents for textmining objects}
\usage{
getDoc(x, i = 1)
}
\arguments{
\item{x}{object to retrive document fe tmCorpus or tmTextDocument}

\item{i}{index}
}
\value{
returns i-th document of object x
}
\description{
Function to access documents for textmining objects
}
\examples{
corp <- tmCorpus(c("This is first document", "This is second"))
getDoc(corp, 1)
getDoc(corp, 2)
content(corp)

text <- tmTextDocument("Text document new")
getDoc(text)
content(text)

parsed <- tmParsed(list(c("Parsed", "doc", "one"), c("Parsed", "two")))
getDoc(parsed, 2)
content(parsed)

}

