% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{make_tabled}
\alias{make_tabled}
\title{Function to create tmWordCountsTable object from tmParsed}
\usage{
make_tabled(x)
}
\arguments{
\item{x}{tmParsed source}
}
\value{
returns tmWordCountsTable object
}
\description{
Function to create tmWordCountsTable object from tmParsed
}
\examples{
corp <- tmCorpus(c("This is first document", "This is second"))
parsed <- parse(corp)
tabled <- make_tabled(parsed)
content(tabled)

}

