% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mallet_prepare}
\alias{mallet_prepare}
\title{Helper function to use mallet topic modelling with tmCorpus}
\usage{
mallet_prepare(doc)
}
\arguments{
\item{doc}{single document}
}
\value{
returns named character vector
}
\description{
Helper function to use mallet topic modelling with tmCorpus
}

