% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{ngram}
\alias{ngram}
\title{Function to create ngram docs}
\usage{
ngram(x, k = 1)
}
\arguments{
\item{x}{tmCorpus object}

\item{k}{number of words in ngram}
}
\value{
returns tmParsed object of ngrams
}
\description{
Function to create ngram docs
}
\examples{
corp <- tmCorpus(c("This is first document", "This is second"))
parsed_ngrams <- ngram(corp, k = 2)
tabled <- make_tabled(parsed_ngrams)
content(tabled)

}

