% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{parse}
\alias{parse}
\title{Function to parse tmCorpus. As an outpus we have tmParsed object.}
\usage{
parse(x)
}
\arguments{
\item{x}{tmCorpus object}
}
\value{
returns tmParsed object
}
\description{
Function to parse tmCorpus. As an outpus we have tmParsed object.
}
\examples{
corp <- tmCorpus(c("This is first document", "This is second"))
parsed <- parse(corp)
content(parsed)

}

