% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{tabler}
\alias{tabler}
\title{Helper function for tabelarising documents}
\usage{
tabler(x)
}
\arguments{
\item{x}{parsed tmTextDocument object.}
}
\value{
returns tmTextDocument with table of words counts
}
\description{
Helper function for tabelarising documents
}

