% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content.R
\name{terms}
\alias{terms}
\alias{terms.jobjRef}
\alias{terms.tmTopicModel}
\title{Function to return the most frequent terms of tmTopicModels}
\usage{
terms(x, ...)

\method{terms}{tmTopicModel}(x, ...)

\method{terms}{jobjRef}(x, ...)
}
\arguments{
\item{x}{tmTopicModel}

\item{...}{addictional parameters fe k - number of terms per topic}
}
\value{
data.frame with terms for topics
}
\description{
Function to return the most frequent terms of tmTopicModels
}

