% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{tmCorpus}
\alias{tmCorpus}
\title{Function to create tmCorpus}
\usage{
tmCorpus(x = NULL, ..., package = "base")
}
\arguments{
\item{x}{source, for package stylo it supposed to be directory with Corpus files}

\item{...}{other arguments dependent on package that we use}

\item{package}{package of reading the data could be tm or stylo}
}
\value{
returns tmCorpus object
}
\description{
Function to create tmCorpus
}
\examples{
corp <- tmCorpus(c("This is corp", "Document 2"))
corp <- tmCorpus(list("This is corp", "Document 2"))
corp <- tmCorpus(VectorSource(c("This is corp", "Document 2")), package = "tm")
\dontrun{
corp <- tmCorpus(DirSource("directory"), package = "tm")
corp <- tmCorpus("directory", package = "stylo")
}

}

