% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{tmMetaData}
\alias{tmMetaData}
\title{Function to create tmMetaData}
\usage{
tmMetaData(id = 1, language = "en", author = character(0),
  date = Sys.Date(), title = paste("Document_", id, sep = ""), ...)
}
\arguments{
\item{id}{id of document}

\item{language}{string language of document}

\item{author}{string authors name}

\item{date}{Date - date of reading in/date of publication}

\item{title}{string title of document}

\item{...}{other metadata}
}
\value{
returns tmMetaData object
}
\description{
Function to create tmMetaData
}
\examples{
tmMetaData(id = 1, author = "Author", newmetadata = "random")
tmMetaData(title = "New title")

}

