% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{tmParsed}
\alias{tmParsed}
\title{Function to create tmParsed}
\usage{
tmParsed(x = NULL, package = "base", ...)
}
\arguments{
\item{x}{source}

\item{package}{package of reading the data could be tm or stylo}

\item{...}{metadata for the tmParsed document}
}
\value{
returns tmParsed object
}
\description{
Function to create tmParsed
}
\examples{
corp <- tmCorpus(c("This is corp", "Document 2"))
parsed <- parse(corp)
parsed_ngram <- ngram(corp, k = 2)
parsed <- tmParsed(list(c("This", "is", "corp"), c("Document", "2")))
\dontrun{
corp <- tmCorpus("directory", package = "stylo")
parsed <- parse(corp)
parsed_ngram <- ngram(corp, k = 2)
parsed <- tmParsed(source = "directory", package = "stylo")
}

}

