% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{tmTaggedCorpus}
\alias{tmTaggedCorpus}
\title{Function to create tmTaggedCorpus}
\usage{
tmTaggedCorpus(x = NULL, ..., treetagger = "manual", lang = "en",
  path = "C:\\\\TreeTagger", preset = "en")
}
\arguments{
\item{x}{source, for package stylo it supposed to be directory with Corpus files}

\item{...}{metadata}

\item{treetagger}{treetagger settings}

\item{lang}{language}

\item{path}{treetagger path}

\item{preset}{language seting}
}
\value{
returns tmTaggedCorpus object
}
\description{
Function to create tmTaggedCorpus
}
\examples{
\dontrun{
corp <- tmCorpus(c("This is corp", "Document 2"))
tg_corp <- tmTaggedCorpus(corp)
}
}

