% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{tmWordCountsTable}
\alias{tmWordCountsTable}
\title{Function to create tmWordCountsTable}
\usage{
tmWordCountsTable(x = NULL)
}
\arguments{
\item{x}{source tmParsed}
}
\value{
returns tmWordCountsTable object
}
\description{
Function to create tmWordCountsTable
}
\examples{
corp <- tmCorpus(c("This is corp", "Document two is this"))
parsed <- parse(corp)
parsed_ngram <- ngram(corp, k = 2)
tabled <- make_tabled(parsed)
tabled_ngram <- make_tabled(parsed_ngram)
\dontrun{
corp <- tmCorpus("directory", package = "stylo")
parsed <- parse(corp)
parsed_ngram <- ngram(corp, k = 2)
parsed <- tmParsed(source = "directory", package = "stylo")
}

}

