% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{topic_network}
\alias{topic_network}
\title{Function to plot topic network}
\usage{
topic_network(k, model)
}
\arguments{
\item{k}{Number of words from each topic to be included in graph}

\item{model}{trained tmTopicModel object}
}
\value{
network The graph visualising the network
}
\description{
Function to plot topic network
}
\examples{
\dontrun{
x <- tmCorpus(rep("as, a , a ,s  l k l l k k j h g f f hg j aaa", 100))
require(rJava)
model <- suppressMessages(train(x))
table_topic <- topic_table(model)
network <- topic_network(10 ,table_topic$words)
}
}

