% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{topic_wordcloud}
\alias{topic_wordcloud}
\title{Simple wordcloud visualization of the topics.}
\usage{
topic_wordcloud(model, topic_id = 1, k = 10, rot_per = 0,
  random_order = FALSE)
}
\arguments{
\item{model}{tmTopicModel object}

\item{topic_id}{Id of the analised topic.}

\item{k}{number of words to be ploted.}

\item{rot_per}{wordcloud param}

\item{random_order}{order of words}
}
\description{
Simple wordcloud visualization of the topics.
}
\examples{
\dontrun{
library(rJava)
x <- tmCorpus(lapply(1:100, function(x) paste(sample(LETTERS, 11),
                                              collapse = "")))

model <- train(x)
topic_wordcloud(model, topic_id = 2, k = 11)
}
}

