% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{train}
\alias{train}
\alias{train.DocumentTermMatrix}
\alias{train.tmCorpus}
\title{train for \code{tmCorpus} object}
\usage{
\method{train}{tmCorpus}(x, k = 20, stoplist_file = "en.txt",
  token_regexp = regexp_token, alpha_opt = 20, burn_in = 50,
  train = 200, maximize = 10, package = "mallet", ...)

\method{train}{DocumentTermMatrix}(x, k = 20, ...)
}
\arguments{
\item{x}{A \code{tmCorpus} object or \code{DocumentTermMatrix} object}

\item{k}{number of topics}

\item{stoplist_file}{file direcroty or vector of stopwords}

\item{token_regexp}{regular expression token}

\item{alpha_opt}{mallet LDA topic model parameter}

\item{burn_in}{mallet LDA topic model parameter}

\item{train}{mallet LDA topic model parameter}

\item{maximize}{mallet LDA topic model parameter}

\item{package}{package to train topic mdoel can be set to "mallet" or "topicmodels"}

\item{...}{other model arguments}
}
\description{
#' Function to create train Topic Model
#'
#' @param x tmCorpus object
#' @param ... settings for mallet.doc.topics and mallet.topic.words
#'
#' @return returns object of a class tmTopicModel
#'
#' @export
train <- function(x, ...) {
  UseMethod("train")
}
}

